Include source log lines in ES error reports
Since we stream documents into ES, keep a LRU cache larger than the ES
chunk size and use this to find the source document when getting an
error document.

Requires source documents to specify _id. This will probably become
mandatory in future (the handler could do it, but it would require
deserialising from JSON. Or would it...?)